% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSHMM-class.R
\docType{methods}
\name{as.character,TSSHMM-method}
\alias{as.character,TSSHMM-method}
\title{Allow TSSHMM to be converted to character vector for logging.}
\usage{
\S4method{as.character}{TSSHMM}(x)
}
\arguments{
\item{x}{TSSHMM model to convert to character vector for logging.}
}
\value{
Character vector of column concatenated transition and emission
matrices.
}
\description{
Allow TSSHMM to be converted to character vector for logging.
}
