% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSHMM-class.R
\docType{import}
\name{coerce}
\alias{coerce}
\alias{coerce,TSSHMM,character-method}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{methods}{\code{\link[methods:setAs]{coerce}}}
}}

