% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{core2014}
\alias{core2014}
\title{Read counts of GRO-seq, and GRO-cap with and without TAP.}
\format{
GRangesList with 3 GRanges objects:
\describe{
\item{GROcap_wTAP:}{read counts, GRO-cap with TAP.}
\item{GROcap_noTAP:}{read counts, GRO-cap without TAP.}
\item{GROseq:}{read counts, GRO-seq.}
}
}
\source{
GSE60453 and GSE60454
}
\usage{
core2014
}
\description{
Dataset of K562 cells from Core et al 2014.  The GRO-cap was differentially
treated with and without the critical 5'-cap removing enzyme, tobacco acid
phosphatase (TAP).
}
\keyword{datasets}
