% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{hmm}
\alias{hmm}
\title{Convert signal and background GRO-cap counts into promoter regions.}
\usage{
hmm(signal, bg, ranges)
}
\arguments{
\item{signal}{Stranded, single base \code{GRanges} with integer score.}

\item{bg}{Stranded, single base \code{GRanges} with integer score.}

\item{ranges}{\code{GRanges} to limit signal search.}
}
\value{
Strand-specific \code{GRanges} of peaked and non-peaked promoter
regions.
}
\description{
\code{hmm} implements Andre Martin's HMM model to find transcription start
sites from GRO-cap sequencing.
}
\details{
The model takes inputs of raw GRO-cap counts, aggregates 10 basepair tiles
by maximum value, and categorizes them into 3 types of observations:
\enumerate{
\item no signal (TAP+ == 0)
\item enriched  (TAP+ >  TAP-)
\item depleted  (TAP- >= TAP+ > 0)
}
...and searches for 3 groups of hidden states:
\enumerate{
\item background,
\item peaked TSS regions, and
\item non-peaked TSS regions
}

Both the peaked and non-peaked TSS regions each require 3 states to describe
them:
\itemize{
\item Non-peaked regions flanked by single low intensity transitions and a
moderately long intensity center.
\item Peaked regions flanked by one-or-more low intensity transitions and a
short intense center.
}
Therefore, in total the HMM has 7 states:
\describe{
\item{B}{Background}
\item{N1}{Non-peaked TSS transition state}
\item{N2}{Non-peaked TSS repeating state}
\item{N3}{Non-peaked TSS transition state}
\item{P1}{Peaked TSS moderate signal}
\item{P2}{Peaked TSS high signal}
\item{P3}{Peaked TSS moderate signal}
}

Finally, after the hidden states are obtained from Viterbi decoding, the
hidden states are converted back into stranded GRanges.
}
\examples{
signal <- GRanges(paste0("chr1:", c(100, 110, 200, 300), ":+"))
score(signal) <- rep(5L, 4)
signal
promoters_peaked <- hmm(signal, GRanges(), range(signal))
promoters_peaked
# There is only 1 promoter in thise region because the first two signal
# values filling 2x 20 bp windows are captured by the HMM as a promoter.
# The remaining 2 signal peaks with no surrounding signal are ignored.
stopifnot(length(promoters) == 1)

bg_nonpeaked <- signal
promoters_nonpeaked <- hmm(signal, bg_nonpeaked, range(signal))
promoters_nonpeaked
}
\references{
\insertRef{core_analysis_2014}{tsshmm}
}
