% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSHMM-class.R
\name{sprint_precision_or_na_spaces}
\alias{sprint_precision_or_na_spaces}
\title{Convert matrix to string vectors, substituting NA values with spaces.}
\usage{
sprint_precision_or_na_spaces(x, collapse = "  ")
}
\arguments{
\item{x}{Matrix which may have NA values.}

\item{collapse}{characters between matrix columns.}
}
\value{
Character vectors of matrix rows with maximum inferred per column.
}
\description{
Convert matrix to string vectors, substituting NA values with spaces.
}
