% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSHMM-class.R
\docType{methods}
\name{coerce}
\alias{coerce}
\alias{coerce,TSSHMM,character-method}
\title{Allow TSSHMM to be converted to character vector for logging.}
\description{
Allow TSSHMM to be converted to character vector for logging.
}
