% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSHMM-class.R
\docType{class}
\name{TSSHMM-class}
\alias{TSSHMM-class}
\title{Transcription start site (TSS) hidden Markov model (HMM).}
\value{
A default model object to be trained and then used for decoding.
}
\description{
Implementation of the TSS HMM model using the General Hidden Markov Model
(GHMM) library.  Initialize a pre-designed hidden Markov model.  Due to the
complexity of the C-interface of the model, the model is managed in the
C-layer and the R layer merely manages the external reference pointer and
provides the print function; there are no setters to arbitrarily modify the
model besides running the training and decoding functions.
}
